+++ Makefile
@@ -16,7 +16,7 @@ GIT_BRANCH=$(shell git rev-parse --abbrev-ref HEAD)
 
 # Build variables
 BUILD_VARS=-X main.GitCommit=${GIT_COMMIT} -X main.GitBranch=${GIT_BRANCH} -X main.BuildTime=${BUILD_TIME} -X main.GitClean=${GIT_CLEAN} -X main.LastGitTag=${LAST_GIT_TAG} -X main.GitTagIsCommit=${GIT_IS_TAG_COMMIT}
-BUILD_FILES=glauth.go bindata.go ldapbackend.go webapi.go configbackend.go sqlitebackend.go localtoolbox.go
+BUILD_FILES=glauth.go bindata.go ldapbackend.go webapi.go configbackend.go sqlitebackend.go mysqlbackend.go localtoolbox.go
 
 #####################
 # High level commands
